Take in charge several copies in different output paths
Previously, only the original path was checked when copying. If a second copy is requested on another output path, it resulted in no copy as the checked considered the original file had already been copied.
Now the check is done based on the original and the output path.
This implies a database structure change as the original path is no more unique.

Refactored also the script to make it cleaner.